updated so that rejected CVEs are correctly removed from the DB per #3148
